Move actions to actionTypes, and refactor all usage
