+++ client/src/actionCreators/voterKey.js
@@ -1,4 +1,4 @@
-import { UPDATE_VOTER_KEY } from '../actions/voterKey';
+import { UPDATE_VOTER_KEY } from '../actionTypes/voterKey';
 
 export const updateVoterKey = key => ({
   type: UPDATE_VOTER_KEY,
