+++ client/src/reducers/votingEnabled.js
@@ -1,4 +1,4 @@
-import { DISABLE_VOTING, ENABLE_VOTING } from '../actions/voting';
+import { DISABLE_VOTING, ENABLE_VOTING } from '../actionTypes/voting';
 
 const votingEnabled = (state = false, action) => {
   switch (action.type) {
